fixed sorting error

+++ src/main/java/com/starter/starter_demo/crud/entity/Orderline.java
@@ -4,8 +4,6 @@ import java.io.Serializable;
 import java.util.Date;
 
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -13,6 +11,7 @@ import javax.persistence.NamedQuery;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name="orderline")
@@ -26,7 +25,7 @@ public class Orderline implements Serializable {
 	private String orderCode;
 	private int status;
 	private String number;
-	private float price;
+	private BigDecimal price;
 	@Temporal(TemporalType.TIMESTAMP)
 	private Date createdAt;
 	@Temporal(TemporalType.TIMESTAMP)
@@ -108,11 +107,11 @@ public class Orderline implements Serializable {
 		this.promo = promo;
 	}
 
-	public float getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
 
-	public void setPrice(float price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
